RobloxScripts/MathBoxingAnimScript/python/requirements.txt
selenium
Pillow
pyperclip
